add binpack and spread policy
