+++ autonomous-scheduler/vGPUScheduler/Makefile
@@ -5,14 +5,14 @@ local:
 
 build:
 
-	sudo docker build --no-cache . -t centaurusinfra/vgpu-scheduler:0.3.0
+	sudo docker build --no-cache . -t centaurusinfra/vgpu-scheduler:0.3.1
 
 push:
-	sudo docker push centaurusinfra/vgpu-scheduler:0.3.0
+	sudo docker push centaurusinfra/vgpu-scheduler:0.3.1
 
 # Run go fmt against code
 fmt:
 	sudo gofmt -l -w .
 
-clean: fmt vet
+clean: fmt
 	sudo rm -f kube-scheduler
