+++ autonomous-scheduler/vGPUScheduler/cmd/scheduler/main.go
@@ -8,7 +8,9 @@ import (
 
 	"k8s.io/component-base/logs"
 	"k8s.io/kubernetes/cmd/kube-scheduler/app"
-	"vGPUScheduler/pkg/vGPUScheduler"
+	// 	"vGPUScheduler/pkg/vGPUScheduler"
+	"vGPUScheduler/pkg/alnair-cost-saving"
+	"vGPUScheduler/pkg/alnair-high-performance"
 
 	_ "sigs.k8s.io/scheduler-plugins/pkg/apis/config/scheme"
 )
@@ -19,7 +21,9 @@ func main() {
 	defer logs.FlushLogs()
 
 	cmd := app.NewSchedulerCommand(
-		app.WithPlugin(vGPUScheduler.Name, vGPUScheduler.New),
+		// 	    app.WithPlugin(vGPUScheduler.Name, vGPUScheduler.New),
+		app.WithPlugin(alnairhighperformance.Name, alnairhighperformance.New),
+		app.WithPlugin(alnaircostsaving.Name, alnaircostsaving.New),
 	)
 
 	if err := cmd.Execute(); err != nil {
