+++ autonomous-scheduler/vGPUScheduler/manifests/kube-scheduler.yaml
@@ -19,7 +19,7 @@ spec:
     - --port=0
     - --config=/etc/kubernetes/manifests/vGPUScheduler-config.yaml
     - --v=5
-    image: centaurusinfra/vgpu-scheduler:0.3.0
+    image: centaurusinfra/vgpu-scheduler:0.3.1
     imagePullPolicy: IfNotPresent
     livenessProbe:
       failureThreshold: 8
@@ -64,26 +64,3 @@ spec:
       path: /etc/kubernetes/manifests/vGPUScheduler-config.yaml
       type: FileOrCreate
 status: {}
----
-# Second part, apply extra privileges to system:kube-scheduler.
-kind: ClusterRole
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: system:kube-scheduler:plugins
-rules:
-- apiGroups: [""]
-  resources: ["pods"]
-  verbs: ["get", "list", "watch", "create", "delete", "update", "patch"]
----
-kind: ClusterRoleBinding
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: system:kube-scheduler:plugins
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: system:kube-scheduler:plugins
-subjects:
-- apiGroup: rbac.authorization.k8s.io
-  kind: User
-  name: system:kube-scheduler
