--- autonomous-scheduler/vGPUScheduler/pod2.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: gpu-pod2
-spec:
-  containers:
-    - name: cuda-container
-      image: nvidia/cuda:9.0-devel
-      resources:
-        limits:
-          alnair/vgpu-memory: 10
-      command: ["sleep", "3600"]
