+++ autonomous-scheduler/vGPUScheduler/test/pod_1.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: gpu-pod-04
+spec:
+#  default-scheduler or
+#  schedulerName: alnair-cost-saving
+  containers:
+    - name: cuda-container
+      image: nvidia/cuda:9.0-devel
+      resources:
+        limits:
+          alnair/vgpu-memory: 4
+      command: ["sleep", "600"]
