+++ autonomous-scheduler/vGPUScheduler/test/pod_2.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: gpu-pod-06
+spec:
+#  schedulerName: alnair-high-performance
+  schedulerName: alnair-cost-saving
+  containers:
+    - name: cuda-container
+      image: nvidia/cuda:9.0-devel
+      resources:
+        limits:
+          alnair/vgpu-memory: 6
+      command: ["sleep", "600"]
